requirements.txt
python-telegram-bot[job-queue]==21.4
pymongo==4.8.0
python-dotenv==1.0.1
google-generativeai==0.7.2
